New separate module for custom errors.
